requirements.txt
streamlit
imap-tools
pandas
anthropic
python-dotenv
tqdm
beautifulsoup4
requests
pypdf
